dropdowns

+++ database/migrations/2023_01_10_125937_create_deductions_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('deductions', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->float('price',8 ,2)->default('0.00');
+            $table->double('price',8 ,2)->default('0.00');
             $table->string('type');
             $table->timestamps();
         });
